Refactor shownPersons and fix name comparison
shownPersons is now an array initialized every time the app renders. Since changing the state of search will trigger the rendering of the whole app, shownPersons will thus be in sync with the value of search. Therefore, there is no need to hold shownPersons in another state. This simplifies the code logic.
The check on whether the person already exists is now changed to case-insensitive.