Finish 5.5 - 5.16
Note that for 5.15 and 5.16, the addBlog and updateLikes function are supposed to be passed in as props. However, when I refactored the code, I think it is better to put them inside the child components inside App since these functions won't be used outside of the child components. But after this change, the exercise 5.15 - 5.16 doesn't apply so well to my code because I will be testing the state modifiers instead of the addBlog and updateLikes function directly. I still followed the exercises' instructions and test against the props passed in (aka the state modifiers) instead, despite that this approach may not make the most sense.